Rename pd-launch scripts to run_slurm{,_cli}.py
